+++ lib/js_of_ocaml/js_of_ocaml.ml
@@ -24,6 +24,7 @@
 module Abort = Abort
 module CSS = CSS
 module Console = Console
+module Crypto = Crypto
 module Dom = Dom
 module Dom_events = Dom_events
 module Dom_html = Dom_html